requirements.txt
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
charset-normalizer==3.3.2
expiring-dict==1.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
ruff==0.6.9
sniffio==1.3.1
sortedcontainers==2.4.0
sseclient-py==1.8.0
typing_extensions==4.12.2
urllib3==2.2.3
